+++ src/include/hashkey.h
@@ -6,8 +6,8 @@
 
 #define HASH_IDX(p) ((p > 8) ? p - 3 : p - 1)
 
-const uint64_t HASHTABLE[12][64];
+const unsigned long long HASHTABLE[12][64];
 
-uint64_t hashkey_position(Position* Position);
+unsigned long long hashkey_position(Position* Position);
 
 #endif
